refactor(cache): use cacheMaxMemorySize config in default cache handler
This refactoring allows the cacheMaxMemorySize configuration to control
the memory limits of the default "use cache" handler, which was
previously hardcoded to 50MB.

Changes:
- Extract cache handler creation into createDefaultCacheHandler() factory
- Thread cacheMaxMemorySize through initializeCacheHandlers()
- Harmonize parameter naming (maxMemoryCacheSize → cacheMaxMemorySize)
- Make cacheMaxMemorySize required in type signatures
- Support zero-size cache (returns no-op handler when size=0)

This enables users to disable internal caching when using external CDNs
or to customize memory limits based on their deployment environment.
