+++ packages/next/src/export/helpers/create-incremental-cache.ts
@@ -20,7 +20,7 @@ export async function createIncrementalCache({
   requestHeaders,
 }: {
   cacheHandler?: string
-  cacheMaxMemorySize?: number
+  cacheMaxMemorySize: number
   fetchCacheKeyPrefix?: string
   distDir: string
   dir: string
@@ -38,7 +38,7 @@ export async function createIncrementalCache({
     )
   }
 
-  if (cacheHandlers && initializeCacheHandlers()) {
+  if (cacheHandlers && initializeCacheHandlers(cacheMaxMemorySize)) {
     for (const [kind, handler] of Object.entries(cacheHandlers)) {
       if (!handler) continue
